back/requirements.txt
Django==2.2.15
django-appconf==1.0.4
django-cors-headers==3.5.0
django-imagekit==4.0.2
django-rest-auth==0.9.5
django-settings-json==1.0.2
djangorestframework==3.11.1
mysqlclient==2.0.1
pilkit==2.0
Pillow==7.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
